kfc-backend/requirements.txt
boto3==1.34.14
aws-xray-sdk==2.12.1
pydantic==2.5.3
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
